fix(python): fix linking an already installed version
Before the fix, the following fails:
> install-tool python 3.7.0 && install-tool python 3.6.0 && install-tool python 3.7.0
with:
> ...
> linking tool python v3.7.0
> No /opt/buildpack/tools/python/3.7.0/bin/python. defined - aborting

The link_tool() function for python uses the MAJOR and MINOR env
variables that are set by the check_semver(), called by
the check_tool_requirements().

The check_tool_requirements() is called only for a fresh install of the
tool, so when linking an already installed version (e.g. to change the
default python version), the check_tool_requirements() call is skipped,
the MAJOR and MINOR env variables are missing and the error is raised.

This change addresses this issue by calling the check_semver() in the
link_tool() for python.