feat(python): Add keyrings.envvars package to pip-tools (#2145)
* feat(pip-tools): Add keyrings.envvars package

Add the keyrings.envvars package to the pip-tools venv to allow passing
credentials to pip-compile via environment variables.

* refactor(python): Accept optional extra packages to install

I'm planning to add the keyrings.envvars package to be installed with
the pip-tools tool, so instead of copy/pasting the common python
install_tool logic a second time this refactors it into a common
function that will install any packages given as arguments along with
the tool being installed.

* Allow for the posibilty of a python 4.x in the future

* Update src/usr/local/containerbase/utils/python.sh

Co-authored-by: Michael Kriese <michael.kriese@visualon.de>

* Add a comment explaining why keyrings.envvars >= 1.1.0 is needed

* Add basic tests for installing the pip-tools tool

* Remove arm64 test

---------

Co-authored-by: Michael Kriese <michael.kriese@visualon.de>